requirements.txt
# VocalGuard Backend Dependencies

# Flask and web framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# OpenAI for scam detection
openai==1.10.0
# Pin httpx to a compatible version; openai 1.10.0 passes the deprecated
# `proxies` argument that was removed in httpx 0.28.
httpx==0.27.2

# HTTP requests for ElevenLabs API
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# WSGI server for production
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
langdetect==1.0.9
